requirements.txt
streamlit
crewai
langchain
langchain-community
chromadb
sentence-transformers
python-dotenv
pandas
openai